refactor(ai): extraer caché del adaptive engine a submódulo (#445)
Mueve la maquinaria de caché (CachedEntry, FnvHasher, cache_key,
cache_insert, cache_len) de adaptive_engine.rs al submódulo
adaptive_engine/cache.rs, siguiendo la convención de módulos con mod.rs.

- Convierte adaptive_engine.rs en adaptive_engine/mod.rs (git mv).
- Agrega cache_get para encapsular get + expiración + clon; los campos
  de CachedEntry quedan privados al submódulo.
- Superficie pública sin cambios: cache_len sigue pub; resto pub(super).
  Sin churn en call sites.

Refactor puro: comportamiento preservado. Tests unitarios (8) y de
integración (6) del adaptive engine pasan sin cambios con la feature
adaptive-selectors.
